requirements.txt
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-environ==0.4.5
google-api-core==1.30.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
packaging==21.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
tmdbsimple==2.1.0

